// src/openapi/client/models/AdvertisementFilterParameterResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Advertisement filter parameter response
 * @export
 * @interface AdvertisementFilterParameterResponse
 */
export interface AdvertisementFilterParameterResponse {
    /**
     * 
     * @type {string}
     * @memberof AdvertisementFilterParameterResponse
     */
    filterName: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementFilterParameterResponse
     */
    filterNameTranslated: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementFilterParameterResponse
     */
    value: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementFilterParameterResponse
     */
    valueTranslated: string;
}

/**
 * Check if a given object implements the AdvertisementFilterParameterResponse interface.
 */
export function instanceOfAdvertisementFilterParameterResponse(value: object): value is AdvertisementFilterParameterResponse {
    if (!('filterName' in value) || value['filterName'] === undefined) return false;
    if (!('filterNameTranslated' in value) || value['filterNameTranslated'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    if (!('valueTranslated' in value) || value['valueTranslated'] === undefined) return false;
    return true;
}

export function AdvertisementFilterParameterResponseFromJSON(json: any): AdvertisementFilterParameterResponse {
    return AdvertisementFilterParameterResponseFromJSONTyped(json, false);
}

export function AdvertisementFilterParameterResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdvertisementFilterParameterResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'filterName': json['filterName'],
        'filterNameTranslated': json['filterNameTranslated'],
        'value': json['value'],
        'valueTranslated': json['valueTranslated'],
    };
}

export function AdvertisementFilterParameterResponseToJSON(json: any): AdvertisementFilterParameterResponse {
    return AdvertisementFilterParameterResponseToJSONTyped(json, false);
}

export function AdvertisementFilterParameterResponseToJSONTyped(value?: AdvertisementFilterParameterResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'filterName': value['filterName'],
        'filterNameTranslated': value['filterNameTranslated'],
        'value': value['value'],
        'valueTranslated': value['valueTranslated'],
    };
}